requirements.txt
aniso8601==1.2.1
apispec==0.25.1
auth0-python==3.1.2
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
colorama==0.3.9
decorator==4.1.2
dnspython==1.15.0
email-validator==1.0.2
factory-boy==2.9.2
fake-factory==9999.9.9
Faker==0.8.1
Flask==0.12.2
Flask-Script==2.0.5
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.13.6
pbr==3.1.1
py==1.4.34
pytest==3.2.1
python-dateutil==2.6.1
python-dotenv==0.6.5
python-editor==1.0.3
pytz==2017.2
PyYAML==3.12
requests==2.18.4
six==1.10.0
Tempita==0.5.2
UkPostcodeParser==1.1.1
urllib3==1.22
webargs==1.8.1
Werkzeug==0.12.2
